Side Entrance: Fix basic functionality
As it stands, the pool cannot accept funds in return the normal way
(via call/transfer) so there is no way to execute a normal flash loan.
This hurts realisim and give a big hint for the solution, as there is
only one payable function in the contract.

This adds a `receive()` function to the pool and tests it during the
initial tests.